biotic/biotic.py
import time
from digi.xbee.devices import XBeeDevice

xbee = XBeeDevice("/dev/tty16", 9600)

# Get the XBee network object from the local XBee.
xnet = xbee.get_network()

# Start the discovery process and wait for it to be over.
xnet.start_discovery_process()
while xnet.is_discovery_running():
    time.sleep(0.5)

# Get the list of the nodes in the network.
nodes = xnet.get_devices()
