pyproject.toml
[tool.poetry]
name = "biotic"
version = "0.1.0"
description = ""
authors = ["Kevin Nordnes <kevinnordnes@proton.me>"]
readme = "README.md"
packages = [{include = "biotic"}]

[tool.poetry.dependencies]
python = "^3.10"
scapy = "^2.5.0"
argparse = "^1.4.0"
python3-nmap = "^1.6.0"
pyshark = "^0.5.3"
pyusb = "^1.2.1"
libusb = "^1.0.26b5"
zigpy-znp = "^0.9.2"
zigpy = "^0.53.1"
zigpy-cc = "^0.5.2"
xbee = "^2.3.2"
digi-xbee = "^1.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
